[None][refactor] simplify Gemma4 config and embedding sharing
Consolidate the Gemma4 compatibility configs in one module and make the target input embedding a direct, explicitly named alias while preserving the assistant embedding tied to its LM head.
